define helper actor on his own file
