+++ src/main/scala/com/pagantis/singer/flows/BatchHttp.scala
@@ -1,19 +1,15 @@
 package com.pagantis.singer.flows
 
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
-
-import akka.actor.{Actor, ActorSystem, Props}
+import akka.actor.{ActorSystem, Props}
 import akka.event.{Logging, LoggingAdapter}
 import akka.http.scaladsl.Http
 import akka.stream.ActorMaterializer
 import akka.stream.scaladsl.{JsonFraming, StreamConverters}
 import com.typesafe.config.ConfigFactory
 
-import scala.concurrent.duration._
+import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, ExecutionContextExecutor}
 import scala.util.Success
-import spray.json._
 
 object BatchHttp extends App {
 
@@ -39,37 +35,15 @@ object BatchHttp extends App {
   val frameLength = config.getInt("flow.frame_length")
 
   // This shutdown sequence was copied from another related issue: https://github.com/akka/akka-http/issues/907#issuecomment-345288919
-  def shutdownSequence =
-    Http().shutdownAllConnectionPools
-
-  // A sideways actor will maintain state
-  class RunningSum extends Actor with akka.actor.ActorLogging {
-    var total: Long = 0
-    val startTime: LocalDateTime = LocalDateTime.now()
-
-    private def logCount(count: Long): Unit = {
-      val currentTime = LocalDateTime.now()
-      val logRecord = JsObject(
-        Map(
-          "requests_processed" -> JsNumber(total),
-          "timestamp" -> JsString(currentTime.format(DateTimeFormatter.ISO_DATE_TIME)),
-        )
-      )
-      log.info(s"${logRecord.compactPrint}")
-    }
-
-    override def receive: Receive = {
-      case increment: Int =>
-        total = total + increment
-        if(total % 5000 == 0) logCount(total)
-
-    }
-
-    override def postStop(): Unit = logCount(total)
-
+  def shutdownSequence = {
+    for {
+      http <- Http().shutdownAllConnectionPools()
+      akka <- system.terminate()
+    } yield akka
   }
-  val counter = system.actorOf(Props[RunningSum], "counter")
 
+
+  val counter = system.actorOf(Props[CountLogger], "counter")
 
   import Request._
 
@@ -100,7 +74,6 @@ object BatchHttp extends App {
 
   Await.ready(flowComputation, Duration.Inf)
   Await.ready(shutdownSequence, Duration.Inf)
-  Await.ready(system.terminate, Duration.Inf)
 
 
   flowComputation.value match {
